# static/swagger_file.yaml
swagger: '2.0'
info:
  title: Library Management API
  description: API for managing books in a library
  version: 1.0.0

basePath: /

paths:
  /books:
    get:
      summary: List all books
      responses:
        200:
          description: A list of books
          schema:
            type: array
            items:
              type: object
              properties:
                title:
                  type: string
                author:
                  type: string
                published_year:
                  type: string
                isbn:
                  type: string
                genre:
                  type: string
    post:
      summary: Add a new book
      parameters:
        - in: body
          name: book
          description: The book to add
          schema:
            type: object
            required:
              - title
              - author
              - published_year
              - isbn
            properties:
              title:
                type: string
              author:
                type: string
              published_year:
                type: string
              isbn:
                type: string
              genre:
                type: string
      responses:
        201:
          description: Book added
  /books/search:
    get:
      summary: Search for books
      parameters:
        - in: query
          name: author
          type: string
          description: Author of the book
        - in: query
          name: year
          type: string
          description: Year of publication
        - in: query
          name: genre
          type: string
          description: Genre of the book
      responses:
        200:
          description: A list of books matching the search criteria
          schema:
            type: array
            items:
              type: object
              properties:
                title:
                  type: string
                author:
                  type: string
                published_year:
                  type: string
                isbn:
                  type: string
                genre:
                  type: string
  /books/{isbn}:
    put:
      summary: Update book details by ISBN
      parameters:
        - in: path
          name: isbn
          required: true
          type: string
          description: ISBN of the book to update
        - in: body
          name: book
          description: The updated book details
          schema:
            type: object
            properties:
              title:
                type: string
              author:
                type: string
              published_year:
                type: string
              genre:
                type: string
      responses:
        200:
          description: Updated book
        404:
          description: Book not found
    delete:
      summary: Delete a book by ISBN
      parameters:
        - in: path
          name: isbn
          required: true
          type: string
          description: ISBN of the book to delete
      responses:
        200:
          description: Book deleted